Add Interfaces project and improve WebSocketHandler
This commit introduces a new Interfaces project, which now houses the IEventSerializer interface. Adjustments in the WebSocketHandler have also been made to use the ToString() method instead of the raw message when sending via WebSocket. Changes in WebSocket test file and a new ConnectionManager class are part of this commit as well.