Add new classes and extensions for managing web socket connections
Added several classes and extensions to manage web socket connections on a more granular level. This includes a Connection Manager for managing connections, an Interface for the Connection Manager, a Connection State Manager for managing the state of each connection, as well as a MessageType class for capturing different message types. The WebSocketExtensions allows messages to be sent asynchronously over the sockets.